language bottomsheet in progress

+++ appModules/ui/language/src/commonMain/kotlin/dev/reprator/accountbook/language/modals/ModalLanguageState.kt
@@ -4,4 +4,4 @@ import androidx.compose.runtime.Immutable
 
 
 @Immutable
-data class ModalStateLanguage(val language: String, val id: Int, val isSelected: Boolean = false)
+data class ModalStateLanguage(val language: String, val id: Long, val isSelected: Boolean = false)
